Move modification of unuploaded files to own method
